KeyguardSimPinView: don't send dummy request to get remaining pin
I'm not sure of the behavior on QCOM RILs, but on Broadcom RIL
(and probably other RILs), this will eat into the user's PIN
remaining tries. Given that the user is only given ~3 tries, this
is a very bad idea.

In addition, if onAttachedToWindow is called multiple times before
the RIL replies, this can completely delete the retry quota leading
straight to the PUK screen.

Change-Id: I339f1b6d79d543ae8930a7bb95acf823fd5ae20e
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>