Make preselected item clickable in intent chooser
Make the text black for consistency with the list, also replicate the
longpress behaviour from the list

Change-Id: Idcc8cce8ee20d8231232d5e57f7062efc9efd12a

ResolverActivity: fix null pointer exception

if alwaysUseOption is true but mAdapter.hasFilteredItem() is not, we try
to set a LongClickListener on a non existant view.

move the assingment of the LongClickListener down to ensure
mAdapter.hasFilteredItem() is true.

Change-Id: I04df845325a9a21f5563292e56728dcc51e91623
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>