Avoid ANR loop because of queued-up ANR runnables
When a persistent app is killed, it will be restarted immediately. If
this app had ANR:d, there may be a bunch of runnables with its
ProcessRecord queued up. Since the new process will have the same
ProcessRecord (but a different PID), ActivityManager will consider it
to have ANR:d before it has even attached.

This will check that the ProcessRecord's current PID is equal to the
PID when we posted the ANR.

Change-Id: I2a5d7db2f587eceefa4cc5388f10133ca707c2be
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>