Keyguard: Remove bouncer when re-registering status bar
When changing themes we restart the keyguard which causes the
StatusBarKeyguardViewManager to recreate the bouncer without
first removing the old bouncer.  This would cause a user to not
be able to correctly enter their pin since the wrong PasswordTextView
would be found via findViewById().  This patch removes the bouncer
view if mBouncer is non-null when registering the status bar.

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/StatusBarKeyguardViewManager.java

Change-Id: I392502bb648f04016d0b39fbf337054ff4ca3a89
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>