base: fix window bounds for left hand navigationbar
Currently in left-hand navigationbar mode all drawerlayouts have an
inconsistent presentation because the actual window size calculation
code only takes into consideration the content view. This problem is not
present in the rest of the navigation modes because of the nature of the
position of the navigation bar. This patch fix the display and content
layouts calculation.  To repro just set navigation mode to hand-left and
access an app with a drawerlayout (xe:the email app) in landscape mode.

Change-Id: I1394f39f4f5e7e6ac50b9558657685e86547acf3
TICKET: CYNGNOS-1748
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>