SysUI: Guard against NPE in SignalClusterView
It is possible for mSC to be nulled out before the Runnable posted
in onStateChanged is run, which will cause an NPE when it does run.
This patch adds a null check to avoid this scenario.

addresses occasional systemui fc at boot resulting in log message:

java.lang.NullPointerException: Attempt to invoke interface method
        'boolean com.android.systemui.statusbar.policy.SecurityController
        .isVpnEnabled()' on a null object reference

Change-Id: Ief660ad97899785f83bdb49fb50c42bbf7416852
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>