Smartbar: Cripple screen pinning confirmation and get right to it
Due to our new task lock hint api an instructional dialog is
no longer needed or desired. Also, it is rather impractical to
keep the dialog but create some kind of "on the fly" dynamic
layout in the dialog. So when the screen pinning command comes
down the pipes, just fire it.

Change-Id: Iaed77e9965aa18509fad9c881220ab836897c3e3
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>